+++ src/ska_sdp_instrumental_calibration/workflow/pipelines/bandpass_polarisation.py
@@ -42,6 +42,9 @@ from ska_sdp_instrumental_calibration.processing_tasks.post_processing import (
 from ska_sdp_instrumental_calibration.workflow.pipeline_config import (
     PipelineConfig,
 )
+from ska_sdp_instrumental_calibration.workflow.utils import (
+    export_gaintable_to_h5parm,
+)
 
 warnings.simplefilter(action="ignore", category=FutureWarning)
 
@@ -173,8 +176,12 @@ def run(pipeline_config) -> None:
     # Output hdf5 file
     logger.info("Running graph and returning calibration solutions")
     gaintable.load()
-    logger.info(f"Writing solutions to {config.hdf5_name}")
-    export_gaintable_to_hdf5([gaintable], config.hdf5_name)
+    if config.h5parm_name is not None:
+        logger.info(f"Writing solutions to {config.h5parm_name}")
+        export_gaintable_to_h5parm(gaintable, config.h5parm_name)
+    if config.hdf5_name is not None:
+        logger.info(f"Writing solutions to {config.hdf5_name}")
+        export_gaintable_to_hdf5(gaintable, config.hdf5_name)
 
     # Convergence checks (noise-free demo version)
     #  - Note that this runs the graph again. I tried assigning both vis